pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llamaindex-chat"
version = "0.1.0"
description = "Chat interface with LlamaIndex + MinerU for multi-source document retrieval"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "llama-index>=0.10.0",
    "llama-index-llms-openai-like>=0.1.0",
    "llama-index-embeddings-huggingface>=0.1.0",
    "llama-index-vector-stores-chroma>=0.1.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.9",
    "websockets>=12.0",
    "httpx>=0.28.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "atlassian-python-api>=3.41.16",
    "chromadb>=0.4.0",
    "magic-pdf>=0.7.0",
    "pypdf>=4.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0",
    "mypy>=1.8",
    "ruff>=0.2.0",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
